delay hashifying urls to be friendlier to google crawler
This is an attempt to make our site work will with the google
crawler.  The crawler appears to run some javascript, and we
have javascript that converts regular url paths into hash fragments,
which makes the crawler effectively ignore them.  This is sad,
since it hides some of our meatiest material.

This leaves that javascript to run at click time rather than
on startup.  I'm not sure if this will be sufficient, but it
feels worth a try.

Alternatives could be providing other routes for the google crawler
to find the pages, such as a sitemap.