add beacon to support site
This adds the helpscout beacon to the support site.
If a single user login is available on api.getgrist.com,
that identity is used.  If multiple logins are available,
one is prefilled arbitrarily, but not marked definitively
as the user's identity (so it will show in the contact form
rather than being hidden).  If one team site is available,
its name is used as the company name.