document how to make anchor links
This updates the row menu docs to mention the new "copy anchor link"
option.  Also mentions it in cell editing section.

Adds a script to check for broken links and fixes some broken
links in main docs.  There are some broken links in the api
docs - will look into it.