redo Custom Widget section (#51)
This updates the Custom Widget section:

 * Adds information about `grist.onRecord` and `grist.onRecords`
 * Drops information about other methods (although they remain supported)
 * Adds a practical example
 * Mentions existence of invoice example

Overall, this refocuses the section on read-only custom views, which
seem a practical and useful application.  Other methods are dropped for
now because they require "full" access to a document even if just
reading from it.  This is embarrassing to explain, and bad practice
to encourage now that there is an alternative for some use cases.