Add a link to Community in top bar, and to Function Reference under Reference section.
The link to the Community mirrors a similar link to the Help Center on
the Community site.

The link to Function Reference is a second link in the navigation menu,
which is awkward (on clicking it, the other nav item will be
highlighted) but the Reference section is another natural place to look
for the function reference.