fix some internal links, and add a script to check them (#173)
There are some internal links that have rusted, and aren't detected
as bad because only the anchor/hash is wrong, so the link reaches
a valid page (just not the right part of that page). This adds
a small script to apply the htmlproofer plugin to detect such
problems, and fixes a batch of them.

The plugin is a bit delicate. Headings in some files are migrated
from one markdown syntax to another to make the plugin happy.
I made a small fork of the plugin to make it deal better with
headers with style information. The plugin is not used by default,
and the fork is not installed by default, so as not to raise the
barrier of installing and using grist-help too much.