Support localizing the menus (#399)
* Support localizing the sections of the left menu

Introduces `help/{en,fr}/sections-translations.yml` which will contain translations for the [sections](https://www.mkdocs.org/dev-guide/themes/#section) (items in the menu that contain children). GitLocalize is able to translate the keys

* Add titles in documents metadata

So the titles of the pages in documents themselves, so they are translated along with their content in Git Localize

* Remove titles from mkdocs.yml, they are now in the docs metadata

Each doc now embeds metadata containing the title.

This way, the titles can be localized when consulting a document.

* Add github workflow to automatically run ./docs.py update-languages

This will automatically add missing entries in sections-translations.yml
files, which contain translations for the sections in the left menu.