snapshots: rewrite the various commands into a single Compose setup
Running a series of Docker commands plus initialisation commands is a
bit daunting. There are several steps to take, each of which could be
subtly wrong. It seems very error-prone to me.

This example condenses all of those steps into just two files and a
single command to run: `docker compose up`. Hopefully this will be a
better explanation and easier to follow.