i18n: fix nav translation issues
- rely on page title to translate nav sub menu items. Added example in
french for a few pages. We stop explicitely defining nav title for pages
in the nav, and just define their md file name, so that the nav item is
the page title. This allows easy translations without having to build a
specific system for nav sub item texts.
- French: fix ai-assistant page not being "assistant.md", making the
french content unavailable