header: use an image as main website heading "text":
in the figma file, the heading is the only part with a different font.
Loading a custom font (with two different weights to match the figma
design) feels a bit much only for the heading. And, customizing the html
rendered would require us to override the whole material theme
header.html partial.

Instead, we use an svg as background image of the heading. The svg
renders the text as if we used the necessary font. And we make sure the
actual text is not visible.

This makes changing the website heading a bit more difficult, but the
benefit is that we dont override unnecessary code and we dont overload
the website with too much fonts loaded.