header: make the website title text clickable
the material theme doesn't make the heading clickable. This is a little
bit annoying.
Enabling a link with JS at load seems like a good compromise to enable
this behavior (we don't really want to rewrite the whole header partial
just for that).
Make sure the link is not seen by keyboard/assistive tools as the logo
anchor is already there for those usages.