"toc in nav": just delete everything, the few remaining code broke stuff
the table of content, that is part of the nav panel on mobile, was
broken with the remaining code. don't really got why just like that but
figured it was easier to just remove all that stuff as I don't think
it'll be used anyway and we can't still retrieve wip code with git
history.